backend/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg2==2.9.11
sqlparse==0.5.4
